change :change  folder tree and architecture
